First working find_next_page_link case
We have a find_next_page_link that works for a nytimes article.  There is a
small unit test for this.

I renamed the regression test, test.py, to regression_test.py, as it is a
little more informative.  I also renamed the data directories used by that test
accordingly.