Rework regression test to run without internet
The pages and data used for the regression tests are now fully self-contained.
Previously, we were using local HTML files, but things like images were still
being referenced out on the internet.  We now make full local copies of the
pages and their prerequisites so that the entire test can run without the
internet.

This required a lot of tweaking, as we now use wget download the page and
prerequisites.