Refactor to improve urlfetch and gen_test code
The way URL fetching worked was confusing.  The mapping from URLs to paths on
disks was being created as paths relative to the test data directory, but the
MockUrlFetch implementation needed paths relative to the root project
directory.  This made everything confusing.  This fixes that by making url_maps
always relative to the test data directory.

This change also allows gen_test to be refactored to be shorter and clearer.