Fix links in regression tests
Previous to this change, we used the wget --convert-links option to create
versions of pages that could be fully displayed locally.  However, this is a
problem for some multi-page articles.  Some sites have the later pages of an
article in a different directory from the original page.  For example, the
first page of an ArsTechnica article could be:

http://www.arstechnica.com/section/article.ars

The second page of this article would be:

http://www.arstechnica.com/section/article.ars/2

Images in the first page would have image links look something like
"../images/header.jpg".  However, in the second page, they would look like
"../../images/header.jpg".

When we bring the elements of the second page into our readable version, which
we place alongside the original version of the first page, the paths to these
resources will be wrong.

Instead of having wget convert the links, we leave the links alone.  As part of
the readability pass, we convert everything to absolute links.  That way, when
we pull in elements from other pages, the links remain correct.  Finally, we
use the URL map that maps all of the prerequisite URLs to local paths,
downloaded by wget, to turn the absolute URLs into local paths for the
presentation of results.