Fix handling of breaks that delineate paragraphs
This replaces some code that was attempting to construct paragraphs within
divs that were possibly delineated by breaks.  That code was choking on some
common cases.  For example, if a div had paragraphs delineated by breaks, links
within the paragraphs would be broken out into their own paragraphs, even
though they are obviously meant to be in-line.

This implements a slightly more sophisticated algorithm that explicitly looks
for double-breaks and constructs paragraphs in the appropriate places.  It also
tries to leave existing block display elements alone.

This adds some unit tests and updated the regression tests for this improved
behavior.

The code is still a bit ugly, but it works.  It will be cleaned up in a future
commit.